requirements.txt
discord.py==2.6.4
requests==2.32.5
beautifulsoup4==4.14.2
python-dotenv==1.2.1
aiohttp==3.10.11
httpx==0.27.2
pytesseract==0.3.13
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
